pyproject.toml
# pyproject.toml
[project]
name = "AgentDSDemo"
version = "0.1.0"
description = "基于DeepSeek的智能体开发演示"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "deepseek>=1.0.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "black>=23.0.0", "flake8>=6.0.0"]
web = ["streamlit>=1.28.0", "gradio>=4.0.0"]

[tool.black]
line-length = 88
target-version = ['py38']